pyproject.toml
[project]
name = "cbt-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "fastapi[standard]>=0.120.0",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.27",
    "pandas>=2.2.2",
    "xlrd>=2.0.1",
    "lxml>=5.2.1",
    "reportlab>=4.0.0",
    "openpyxl>=3.1.0",
    "psycopg2-binary>=2.9.11",
]
